auth requirements add auth requirements to routes

+++ routes/api.php
@@ -14,7 +14,7 @@ Route::prefix("v1")->group(function () {
         if (config("app.debug")) {
             Route::get("all/", [VideoController::class, "getAllPaginated"])
             ->middleware(AuthMiddleware::class . ":admin");
-            Route::post("recut/", [VideoController::class, "recut"]);
+            Route::post("recut/", [VideoController::class, "recut"])->middleware(AuthMiddleware::class . ":admin");
 
 
             Route::get('/swagger/swagger.json', function () {
@@ -35,16 +35,16 @@ Route::prefix("v1")->group(function () {
 
         Route::get("/{video_id}/clips", [ClipsController::class, "get_clips"]);
 
-        Route::post("/video", [VideoController::class, "create"]);
-        Route::post("/{video_id}/upload-subs", [VideoController::class, "load_subs"]);
-        Route::post("/{video_id}/upload-video", [VideoController::class, "load_video"]);
-        Route::delete("/{video_id}", [VideoController::class, "delete"]);
+        Route::post("/video", [VideoController::class, "create"])->middleware(AuthMiddleware::class . ":admin");
+        Route::post("/{video_id}/upload-subs", [VideoController::class, "load_subs"])->middleware(AuthMiddleware::class . ":admin");
+        Route::post("/{video_id}/upload-video", [VideoController::class, "load_video"])->middleware(AuthMiddleware::class . ":admin");
+        Route::delete("/{video_id}", [VideoController::class, "delete"])->middleware(AuthMiddleware::class . ":admin");
         Route::get("/{video_id}", [VideoController::class, "get_video"]);
-        Route::get("/{video_id}/logs", [VideoController::class, "get_video_logs"]);
+        Route::get("/{video_id}/logs", [VideoController::class, "get_video_logs"])->middleware(AuthMiddleware::class . ":admin");
 
         Route::get("/{video_id}/clip-by-timing", [ClipsController::class, "getClipByTiming"]);
 
-        Route::post("/{video_id}/recut-video", [VideoController::class, "recut"]);
+        Route::post("/{video_id}/recut-video", [VideoController::class, "recut"])->middleware(AuthMiddleware::class . ":admin");
 
 
         // Route::get("/{video_id}/clips-paginated", [ClipsController::class, "getClipPaginated"]);
